90% Finished with Codebase Changed (Is functional at this point
Several improvements to the codebase, including changes to the response embed and even more adjustments to the code:
Added documentation to all defined functions.
Simplified the embed to only show the player count when there are players online.
With minecraft servers, the bedrock server is no longer pinged since it is the same server as the java address.

The only thing left to do is add more comments and little improvements that I will come across in doing that.